Refactor application into modules
